add a variable to each socket
in order to identify whether that socket is in a game already as a player or spectator.
When a socket emits one of the home page events, check that var. If the socket is already in a game, ignore the event.
This is all part of an effort to be distrustful of the client messages. Only do something about client messages if they are received when expected and have the expected data.
Eg. If a client sends a message saying "I played x card", make sure they had that card in their hand and that it was time for them to play a card before accepting the message.